+++ Firmware/RoboTricklerUI/ui.c
@@ -69,7 +69,7 @@ lv_obj_t *ui_LabelMessages;
 void ui_init( void )
 {
 lv_disp_t *dispp = lv_disp_get_default();
-lv_theme_t *theme = lv_theme_default_init(dispp, lv_palette_main(LV_PALETTE_BLUE), lv_palette_main(LV_PALETTE_RED), true, LV_FONT_DEFAULT);
+lv_theme_t *theme = lv_theme_default_init(dispp, lv_palette_darken(LV_PALETTE_BLUE, 2), lv_palette_main(LV_PALETTE_RED), true, LV_FONT_DEFAULT);
 lv_disp_set_theme(dispp, theme);
 ui_Screen1_screen_init();
 lv_disp_load_scr( ui_Screen1);